Configures self-hosted GitHub Action runners
Adds the `runner` role to configure self-hosted GitHub Action runners.

This role automates the process of downloading, installing, and configuring the runner agent on the target machine. It also manages the runner as a service.

Removes deprecated tasks and associated files.

The runner version and checksum are defined as variables.