Adds monitoring tag and Grafana repo update
Adds a monitoring tag to hosts and configures Alloy to include it as an external label. This enables filtering in Grafana/Loki based on host type.
Additionally, it updates the method to add Grafana APT repository for improved compatibility.

Removes secret files from repo as they are not meant to be tracked by git.