Adds k8s nodes to inventory [skip-ci]
Adds k8s nodes to the Ansible inventory and assigns them to the k8s group.

Moves the monitoring tag from host_vars to group_vars and creates a new group_vars file for k8s nodes. This ensures that the monitoring tag is applied consistently to all nodes within the same group.